move to vitest instead of jest to simplify everything
