+++ apps/demo/vite.config.ts
@@ -4,4 +4,7 @@ import react from '@vitejs/plugin-react';
 // https://vitejs.dev/config/
 export default defineConfig({
     plugins: [react()],
+    build: {
+        chunkSizeWarningLimit: 2000,
+    },
 });
