+++ components/js-api-client/package.json
@@ -11,8 +11,7 @@
     "scripts": {
         "watch": "tsc -W",
         "build": "rm -rf dist && tsup src/index.ts --format esm,cjs --dts",
-        "test": "jest",
-        "test:watch": "jest --watch"
+        "test": "vitest run"
     },
     "repository": {
         "type": "git",
@@ -33,7 +32,7 @@
         "@types/node": "^20.10.9",
         "tsup": "^8.1.0",
         "typescript": "^5.5.2",
-        "jest": "^29.7.0"
+        "vitest": "^1.6.0"
     },
     "dependencies": {
         "dotenv": "^16.4.5",
