+++ components/js-api-client/tests/hydrate.paths.test.ts
@@ -1,4 +1,5 @@
-const { createClient, createProductHydrater } = require('../dist/index.js');
+import { test, expect } from 'vitest';
+import { createClient, createProductHydrater } from '../src';
 
 test('Hydrate Paths', async () => {
     const CrystallizeClient = createClient({
