+++ components/js-api-client/tests/util.ts
@@ -1,10 +1,8 @@
-function walkTree(tree, cb) {
+export const walkTree = (tree: any[], cb: any) => {
     tree.forEach((node) => {
         cb(node);
         if (node?.children?.length) {
             walkTree(node.children, cb);
         }
     });
-}
-
-module.exports = { walkTree };
+};
