+++ components/js-storefrontaware-utils/package.json
@@ -10,7 +10,7 @@
     "scripts": {
         "watch": "tsc -W",
         "build": "rm -rf dist && tsup src/index.ts --format esm,cjs --dts",
-        "test": "jest"
+        "test": "vitest run"
     },
     "repository": {
         "type": "git",
@@ -30,8 +30,8 @@
         "@tsconfig/node20": "^20.1.4",
         "@types/node": "^20.14.9",
         "typescript": "^5.5.2",
-        "tsup": "^8.1.0",
-        "jest": "^29.7.0"
+        "vitest": "^1.6.0",
+        "tsup": "^8.1.0"
     },
     "dependencies": {
         "json-to-graphql-query": "^2.2.5",
