+++ components/node-service-api-request-handlers/package.json
@@ -11,7 +11,7 @@
     "scripts": {
         "watch": "tsc -W",
         "build": "rm -rf dist && tsup src/index.ts --format esm,cjs --dts",
-        "test": "jest"
+        "test": "vitest run"
     },
     "repository": {
         "type": "git",
@@ -32,7 +32,7 @@
         "@types/jsonwebtoken": "^9.0.6",
         "@types/koa": "^2.15.0",
         "@types/node": "^20.14.9",
-        "jest": "^29.7.0",
+        "vitest": "^1.6.0",
         "stripe": "^10.13.0",
         "tsup": "^8.1.0",
         "zod": "^3.23.8"
