+++ components/node-service-api-router/package.json
@@ -9,7 +9,7 @@
     "scripts": {
         "watch": "tsc -W",
         "build": "rm -rf dist && tsup src/index.ts --format esm,cjs --dts",
-        "test": "jest"
+        "test": "vitest run"
     },
     "repository": {
         "type": "git",
@@ -36,7 +36,7 @@
         "@types/koa__cors": "^5.0.0",
         "@types/node": "^20.14.9",
         "tsup": "^8.1.0",
-        "jest": "^29.7.0",
+        "vitest": "^1.6.0",
         "koa": "^2.15.3",
         "zod": "^3.23.8"
     },
