+++ components/reactjs-components/package.json
@@ -13,7 +13,7 @@
     "scripts": {
         "watch": "tsc -W",
         "build": "rm -rf dist && tsup src/index.ts --format esm,cjs --dts",
-        "test": "jest"
+        "test": "vitest run"
     },
     "repository": {
         "type": "git",
@@ -42,8 +42,8 @@
         "@types/node": "^20.14.9",
         "@types/react": "^17 || ^18",
         "@types/react-dom": "^17 || ^18",
-        "jest": "^29.7.0",
         "tsup": "^8.1.0",
+        "vitest": "^1.6.0",
         "typescript": "^5.5.2"
     }
 }
