+++ components/reactjs-hooks/package.json
@@ -10,7 +10,7 @@
     "scripts": {
         "watch": "tsc -W",
         "build": "rm -rf dist && tsup src/index.ts --format esm,cjs --dts",
-        "test": "jest"
+        "test": "vitest run"
     },
     "repository": {
         "type": "git",
@@ -33,7 +33,7 @@
         "@types/react-dom": "^17 || ^18",
         "tsup": "^8.1.0",
         "typescript": "^5.5.2",
-        "jest": "^29.7.0"
+        "vitest": "^1.6.0"
     },
     "dependencies": {
         "@crystallize/js-api-client": "workspace:*"
