+++ components/reactjs-hooks/tests/dummy.test.ts
@@ -1,3 +1,4 @@
+import { test, expect } from 'vitest';
 test('Not implemented', async () => {
     expect(1).toBe(1);
 });
