--- components/schema/jest.config.cjs
@@ -1,9 +1,0 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
-    preset: 'ts-jest',
-    testEnvironment: 'node',
-    modulePathIgnorePatterns: ['dist'],
-    moduleNameMapper: {
-        '^(.*)\\.js$': '$1',
-    },
-};
