+++ components/schema/package.json
@@ -15,21 +15,17 @@
     "scripts": {
         "build": "rm -rf dist && tsup src/index.ts --format esm,cjs --dts",
         "watch": "tsc -W",
-        "test": "jest",
-        "test:watch": "jest --watch"
+        "test": "vitest run"
     },
     "dependencies": {
         "zod": "^3.23.8"
     },
     "devDependencies": {
         "@tsconfig/node20": "^20.1.4",
-        "@types/jest": "^29.5.12",
         "@types/node": "^20.14.9",
-        "jest": "^29.7.0",
         "json": "^11.0.0",
-        "ts-jest": "^29.1.5",
-        "ts-node": "^10.9.2",
         "tsup": "^8.1.0",
-        "typescript": "^5.5.2"
+        "typescript": "^5.5.2",
+        "vitest": "^1.6.0"
     }
 }
