+++ components/schema/tests/shape/itemRelations.test.ts
@@ -1,5 +1,6 @@
 import { ZodError } from 'zod';
 import { ItemRelationsComponentConfig, ItemRelationsComponentConfigSchema } from '../../src/shape/index.js';
+import { expect, test } from 'vitest';
 
 interface testCase {
     name: string;
